requirements.txt
Flask==3.0.3
gunicorn==22.0.0
pandas==2.2.2
numpy==1.26.4
plotly==5.24.1
kaleido==0.2.1
matplotlib==3.9.0
seaborn==0.13.2
openpyxl==3.1.5
google-generativeai
